+++ NEWS.md
@@ -1,3 +1,8 @@
+# favawesome (development version)
+
+* Added new function `use_favawesome_standalone()` to allow app developers to
+  take advantage of this package without adding a dependency to this package.
+
 # favawesome 0.1.0
 
 * Initial CRAN submission.